[not verified] move ai selecto model control to share
