+++ projects/plugins/jetpack/extensions/shared/components/ai-select-model-control/style.scss
@@ -1,0 +1,15 @@
+.jetpack-ai-generate-excerpt-control {
+	.jetpack-ai-generate-excerpt-control__header {
+		.components-base-control__field {
+			display: flex;
+			justify-content: space-between;
+		}
+		.components-base-control__label {
+			margin-bottom: 0;
+		}
+	}
+
+	.components-base-control {
+		margin-bottom: 12px;
+	}
+}
